+++ src/utils/middleware.go
@@ -143,7 +143,13 @@ func SetSecurityHeaders(next http.Handler) http.Handler {
 		w.Header().Set("X-Content-Type-Options", "nosniff")
 		w.Header().Set("X-XSS-Protection", "1; mode=block")
 		w.Header().Set("Content-Security-Policy", "frame-ancestors 'self'")
-		
+				
+		next.ServeHTTP(w, r)
+	})
+}
+
+func SetCosmosHeader(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {		
 		w.Header().Set("X-Served-By-Cosmos", "1")
 		
 		next.ServeHTTP(w, r)